Web-DataBase-SemiconductorStudio/python-hub/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
py-bcrypt==0.4